Fix warning related to calling getModuleDir with only one argument.
I tested this in Magento EE 1.12 but I assume it will work in others. 
The method Mage::getModuleDir expects 2 arguments:
public static function getModuleDir($type, $moduleName)
    {
        return self::getConfig()->getModuleDir($type, $moduleName);
    }
A good reference on this:
http://alanstorm.com/magento_base_directories

Unfortunately it doesn't recognize "Helper" as an argument so I just use an empty string and then prepend Helper before the filename.