Add connection checks for APIs
Performs a basic check on the Zendesk API and the custom Magento API, to
allow admins to easily determine if they are working as expected.

As the Magento API needs to be tested from an external location, an AJAX
request is sent from the browser to determine if the API is accessible.